Food API implementation
Began development for the user to be able to submit queries into the FoodData Central API, very much a skeleton for now. Created corrisponding classes that will be able to hold the information for the GET request.

FIXME:
Need to fully implement the food search. Right now it has a good amount of errors so that will need some debugging and some time.

Once we're able to make successful calls and get the information we should be all good. Here is some documentation I've been looking at to help me with this process

https://app.swaggerhub.com/apis/fdcnal/food-data_central_api/1.0.1#/servers

to look at the table schemas so that we can get the labelNutrients field from the BrandedFoodItem Schema